LayoutTests/imported/w3c/web-platform-tests/html/canvas/offscreen/fill-and-stroke-styles/2d.fillStyle.parse.rgba-eof-expected.txt

2d.fillStyle.parse.rgba-eof


FAIL OffscreenCanvas test: 2d.fillStyle.parse.rgba-eof assert_equals: Red channel of the pixel at (50, 25) expected 0 but got 255
